feature(cost): separate cost evalution and solution searching
